addMember 모듈화 / 오류가 왜 뜨는지 모르겠음
